Mantra checks reveal need to redo kwity New issues, TODOs, incomplete version, and broken
WARNING: this leaves a broken file (mantra.sh) so use the TDD.sh instead

**New issues**:
1. git mail map presence check
2. git attribute
3. run instructions: decided by user, through grepping for patterns and
asking if we need to issue anything

**Rename**
TDD.sh renamed to mantra.sh, though for now TDD.sh removal is NOT
committed since the file is now broken

**TODOs**:
1. added rdz file with what needs to be done
2. added TODOs in mantra.sh to see where and what

This is WIP, but it changes enough to push it. It's broken, but it's
time to consider merges, so...

Software design is hard and in my experience the way to get better at is through practice.
    -- Bart Bakker @ blog.thesoftwarecraft.com